tests: Adapt tests to the new client

+++ algoliasearch/check_logs.go
@@ -4,8 +4,8 @@ func checkGetLogs(params map[string]interface{}) error {
 	for k, v := range params {
 		switch k {
 		case "length", "offset":
-			if _, ok := v.(int64); !ok {
-				return invalidType(k, "int64")
+			if _, ok := v.(int); !ok {
+				return invalidType(k, "int")
 			}
 
 		case "type":
